+++ src/Cocorico/CoreBundle/Validator/Constraints/BookingValidator.php
@@ -29,8 +29,8 @@ class BookingValidator extends ConstraintValidator
 {
     private $bookingManager;
     private $session;
-    private $minStartDelay;
     private $minStartTimeDelay;
+    private $acceptationDelay;
     private $currency;
     private $currencySymbol;
     private $timezone;
@@ -38,23 +38,23 @@ class BookingValidator extends ConstraintValidator
     /**
      * @param BookingManager $bookingManager
      * @param Session        $session
-     * @param int            $minStartDelay
      * @param int            $minStartTimeDelay
+     * @param int            $acceptationDelay
      * @param string         $currency
      */
     public function __construct(
         BookingManager $bookingManager,
         Session $session,
-        $minStartDelay,
         $minStartTimeDelay,
+        $acceptationDelay,
         $currency
     )
     {
         $this->bookingManager = $bookingManager;
         $this->session = $session;
-        $this->minStartDelay = $minStartDelay;
         $this->timezone = $this->session->get('timezone');
         $this->minStartTimeDelay = $minStartTimeDelay;
+        $this->acceptationDelay = $acceptationDelay;
         $this->currency = $currency;
         $this->currencySymbol = Intl::getCurrencyBundle()->getCurrencySymbol($currency);
     }
@@ -107,13 +107,25 @@ class BookingValidator extends ConstraintValidator
         if (in_array('date_range.invalid.min_start', $errors)) {
             $minStart = new DateTime();
             $minStart->setTimezone(new DateTimeZone($this->timezone));
-            if ($this->minStartDelay > 0) {
-                $minStart->add(new DateInterval('P'.$this->minStartDelay.'D'));
+            if ($this->minStartTimeDelay > 0) {
+                $minStart->add(new DateInterval('PT'.$this->minStartTimeDelay.'M'));
             }
             $violations[] = array(
                 'message' => 'date_range.invalid.min_start {{ min_start_day }}',
                 'parameter' => array('min_start_day' => $minStart->format('d/m/Y')),
                 'domain' => 'cocorico'
+            );
+        }
+
+        if (in_array('date_range.invalid.acceptation', $errors)) {
+            $maxAcceptableDate = new DateTime();
+            $maxAcceptableDate->setTimezone(new DateTimeZone($this->timezone));
+            $maxAcceptableDate->add(new DateInterval('PT'.$this->acceptationDelay.'M'));
+            $maxAcceptableDate->add(new DateInterval('P1D'));
+            $violations[] = array(
+                'message' => 'date_range.invalid.min_start {{ min_start_day }}',
+                'parameter' => array('min_start_day' => $maxAcceptableDate->format('d/m/Y')),
+                'domain' => 'cocorico',
             );
         }