+++ src/Cocorico/TimeBundle/Form/Type/DateRangeType.php
@@ -146,7 +146,7 @@ class DateRangeType extends AbstractType
                 'allow_single_day' => true,
                 'end_day_included' => true,
                 'display_mode' => 'range',
-                'min_start_delay' => 0,
+                'min_start_time_delay' => 0,
                 'days_max' => $this->daysMax,
                 'allow_end_in_past' => false,
             )
@@ -183,7 +183,7 @@ class DateRangeType extends AbstractType
                             new OptionsResolver(), array(
                                 'required' => $options["required"],
                                 'allow_single_day' => $options["allow_single_day"],
-                                'min_start_delay' => $options["min_start_delay"],
+                                'min_start_time_delay' => $options["min_start_time_delay"],
                                 'days_max' => $options["days_max"],
                                 'allow_end_in_past' => $options["allow_end_in_past"]
                             )